+++ src/commands/general/index.ts
@@ -3,5 +3,5 @@ export { Help } from './Help';
 export { Info } from './Info';
 export { Invite } from './Invite';
 export { Ping } from './Ping';
-export { Remind } from './Remind';
 export { Stats } from './Stats';
+